+++ CHANGELOG.md
@@ -18,6 +18,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Renamed `debug` package to `fxdebug`
 - Renamed `debug/metrics/prometheus` package to `fxpromhttp`
 - Renamed `dev` package to `fxdev`
+- Renamed `errors` package to `fxerrors`
+- Renamed `errors/airbrake` package to `fxairbrake`
 
 
 ## [0.23.0] - 2018-04-29